apps/document_processor_function/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
openai>=1.0.0
python-dotenv>=1.0.1
azure-core>=1.30.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.24.0
pypdf>=5.4.0
defusedxml>=0.7.1
